Handle exceptions in AddUserModal
AddUserModal was not equipped to handle non-existent users in the previous commit, but now it checks if there was any data to be retrieved and sets the error message to 'No such user found'.
Since the modal is intended to facilitate multiple additions, adding a user now changes the error to 'User added!', and doesn't close the modal until the Cancel button is pressed.