Refactor and add checks to enable LeaveGroupModal
Add useEffect and useState calls to GroupPage to determine if a user
should be allowed to leave.
Namely: the user must be alone in the group, or there is another admin
in the group.

Fix a bug in groups.js that was stopping groups from being deleted
from the firebase side. However, the issue with subcollections not
being deleted will still be present, since those have to be deleted
manually.